refactor(common, salvage): extract title
